+++ tools/redis/samples/types/zset.sh
@@ -18,9 +18,17 @@ zpopmax() {
 zincrby() {
 	redis-cli ZINCRBY ${key} $@
 }
+zcard() {
+	redis-cli ZCARD ${key}
+}
+zscore() {
+	redis-cli ZSCORE ${key}
+}
+
 zadd 1 rongtao1
 zadd 2 rongtao2
 zadd 3 rongtao3
+zscore rongtao3
 
 zrange 0 10
 
@@ -30,3 +38,7 @@ zrange 0 10
 
 zpopmin
 zpopmax
+
+# List all
+zrange 0 -1
+zcard